--- Sydney.Core/RoutingToo/Router.cs
@@ -1,194 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-using Sydney.Core.Handlers;
-using Sydney.Core.Routing;
-
-namespace Sydney.Core.RoutingToo;
-
-internal class Router
-{
-    public Router()
-    {
-        this.root = new PathNode(PathNodeType.Segment, string.Empty);
-        this.handlerPaths = new List<string>();
-    }
-
-    private readonly PathNode root;
-    private readonly List<string> handlerPaths;
-
-    /// <summary>
-    /// Gets the list of registered handler paths. This is used exclusively to display routes when
-    /// the service starts.
-    /// </summary>
-    public IReadOnlyList<string> HandlerPaths
-    {
-        get { return this.handlerPaths; }
-    }
-
-    public void AddHandler(string path, SydneyHandlerBase handler)
-    {
-        // "books//foo/"
-        // "/"
-        string trimmedPath = TrimSlashes(path);
-
-        string[] segments = trimmedPath.Split('/');
-        HashSet<string> parameterNames = new HashSet<string>();
-        PathNode node = this.root;
-        for (int i = 0; i < segments.Length; i++)
-        {
-            string segment = segments[i];
-
-            // The first segment can only be empty if the string is a single slash.
-            if (i > 0 && string.IsNullOrWhiteSpace(segment))
-            {
-                throw new InvalidOperationException(
-                    "Route segments must be at least one non-whitespace character long.");
-            }
-
-            if (TryGetParameterName(segment, out string? parameterName))
-            {
-                if (parameterNames.Contains(parameterName))
-                {
-                    throw new InvalidOperationException(
-                        $"The parameter {parameterName} is used twice in this route. Parameters must be unique.");
-                }
-
-                parameterNames.Add(parameterName);
-                if (node.Parameter != null)
-                {
-                    if (parameterName != node.Parameter.Value.Key)
-                    {
-                        throw new InvalidOperationException(
-                            $"The parameter {parameterName} is attempting to rename an already existing parameter.");
-                    }
-
-                    node = node.Parameter.Value.Value;
-                }
-                else
-                {
-                    PathNode child = new PathNode(PathNodeType.Parameter, parameterName, node);
-                    node.Parameter = new KeyValuePair<string, PathNode>(
-                        parameterName,
-                        child);
-                    node = child;
-                }
-            }
-            else
-            {
-                if (!node.Children.TryGetValue(segment, out PathNode? child))
-                {
-                    child = new PathNode(PathNodeType.Segment, segment, node);
-                    node.Children[segment] = child;
-                }
-
-                node = child;
-            }
-        }
-
-        node.Handler = handler;
-        handlerPaths.Add(trimmedPath);
-    }
-
-    public bool TryMatchPath(
-        string path,
-        [NotNullWhen(returnValue: true)] out MatchResult? matchResult)
-    {
-        List<MatchResult> results = new List<MatchResult>();
-        matchResult = null;
-        string trimmedPath = TrimSlashes(path);
-        string[] segments = trimmedPath.Split('/');
-
-        MatchPathRecursive(
-            this.root,
-            segments,
-            0,
-            new Dictionary<string, string>(),
-            results);
-        matchResult = results.OrderBy((r) => r.PathParameters.Count).FirstOrDefault();
-
-        return matchResult != null;
-    }
-
-    private static void MatchPathRecursive(
-        PathNode node,
-        string[] segments,
-        int index,
-        Dictionary<string, string> pathParametersSoFar,
-        List<MatchResult> results)
-    {
-        string segment = segments[index];
-        if (index == segments.Length - 1)
-        {
-            if (node.Handler != null)
-            {
-                if (node.Type == PathNodeType.Parameter)
-                {
-                    pathParametersSoFar.Add(node.Value, segment);
-                    results.Add(
-                        new MatchResult(
-                            node.Handler,
-                            new Dictionary<string, string>(pathParametersSoFar)));
-                    pathParametersSoFar.Remove(node.Value);
-                }
-                else if (node.Value == segment)
-                {
-                    results.Add(
-                        new MatchResult(
-                            node.Handler,
-                            new Dictionary<string, string>(pathParametersSoFar)));
-                }
-            }
-
-            return;
-        }
-
-        if (node.Children.TryGetValue(segment, out PathNode? child))
-        {
-            MatchPathRecursive(
-                child,
-                segments,
-                index + 1,
-                pathParametersSoFar,
-                results);
-        }
-
-        if (node.Parameter != null)
-        {
-            pathParametersSoFar.Add(node.Value, segment);
-            MatchPathRecursive(
-                node.Parameter.Value.Value,
-                segments,
-                index + 1,
-                pathParametersSoFar,
-                results);
-            pathParametersSoFar.Remove(node.Value);
-        }
-    }
-
-    private static bool TryGetParameterName(
-        string segment,
-        [NotNullWhen(returnValue: true)] out string? parameterName)
-    {
-        if (segment[0] == '{' && segment[segment.Length - 1] == '}')
-        {
-            parameterName = segment.Substring(1, segment.Length - 2);
-            return true;
-        }
-
-        parameterName = null;
-
-        return false;
-    }
-
-    private static string TrimSlashes(string route)
-    {
-        // Trim leading and trailing slashes from the route.
-        string result = route.Trim('/');
-        if (result == "/")
-        {
-            // If the route was just a single slash, Trim won't have removed it.
-            return string.Empty;
-        }
-
-        return result;
-    }
-}
